Remove the test-only HTTP transport, keep in-memory only
ENG-1033 ships one production transport: the in-memory one. The test-only
HTTP adapter and its local/HTTP parity duplication were speculative scaffolding
for a future network transport that does not exist yet, so drop them and keep
the smallest change that satisfies the shipped requirement.

Deletes test/http-transport.ts and its dedicated hostile-wire-peer test file.
parity.test.ts (renamed to call-behavior.test.ts) drops the now-pointless
multi-transport harness and keeps its in-memory test cases. dispatch.ts
un-exports the internal pieces (resolveInterModuleInput,
invokeHandlerWithCancellation, drainReport, safeParseWithDefectDetection,
and their result types) that were only exported for the HTTP adapter to
compose; only runInterModuleCall, the actual transport-authoring seam
transport.ts uses, stays exported. tracing.ts drops the unused
startInterModuleSpan (CLIENT/SERVER span kinds had no remaining caller).

No behavior change to the shipped in-memory transport.
